requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
redis
httpx
pytest
pytest-mock
pytest-dotenv